requirements.txt
pandas
openpyxl
numpy
streamlit
plotly
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
firebase-admin
python-dotenv
